requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.1
fake-useragent==1.4.0
python-dotenv==1.0.1
undetected-chromedriver==3.5.5
tqdm==4.66.2
retry==0.9.2 